refactor: move HomeSection and HomeSectionHeader to common/ui
