+++ apps/frontend/src/common/ui/section/Section.vue
