+++ apps/frontend/src/common/ui/section/SectionHeader.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
 import { RouterLink } from "vue-router";
-import { useCategories } from "@/features/categories/categories.queries";
-import CategoriesGrid from "@/features/categories/views/CategoriesGrid.vue";
 
-const {
-  data: categories,
-  isLoading,
-  error,
-} = useCategories({ sort: "-recipeCount" });
+const props = defineProps<{
+  title: string;
+  subtitle: string;
+  link?: {
+    text: string;
+    to: string;
+  };
+}>();
 </script>
 
 <template>
@@ -16,22 +17,22 @@ const {
       <p
         class="text-terracotta mb-2 text-sm font-semibold tracking-widest uppercase"
       >
-        Pick a direction
+        {{ title }}
       </p>
       <h2
         class="font-display text-4xl font-bold tracking-tight text-stone-900 lg:text-5xl"
       >
-        Recipe Categories
+        {{ subtitle }}
       </h2>
     </div>
+
     <RouterLink
-      to="#"
+      v-if="link"
+      :to="link.to"
       class="text-terracotta hover:text-terracotta-dark hidden items-center gap-2 text-sm font-semibold transition-colors sm:flex"
     >
-      All categories
+      {{ link.text }}
       <i class="pi pi-arrow-right text-xs" />
     </RouterLink>
   </div>
-
-  <CategoriesGrid :categories :isLoading :error />
 </template>
